fix(ci): the Linux build and clippy
`MOE_K`, `MOE_ALPHA` and `MOE_ETA` were declared inside the Metal-only module
of `cortiq-embryo::model`, but the exporter and the growth records read them
on every platform — so the workspace only built on macOS. They are shape
constants, not device ones; they now live outside it.

Plus what a newer clippy flags: a lazily-continued doc list, and the
needless borrows `--fix` found.

And a process-wide `gpu::pause_gpu()`: `cpu_scope` is thread-local and the
pool's workers do not inherit it, so a caller that wants a whole *phase* off
the device has no way to say so. The LTX prompt encoder uses it on a machine
that cannot keep the whole container wired — its weights live in a part of
the file the denoising loop never touches, and putting both on the driver's
books makes it evict between commits while the hot loop pays for it.

Metal windows over the file mapping are now built on first use rather than up
front, for the same reason: materializing every window of a 21 GB file puts
22 GB on the books of a device whose recommended working set is 17.8 GB.

+++ crates/cortiq-cli/src/avout.rs
@@ -179,7 +179,7 @@ impl BitWriter {
     fn flush(&mut self) {
         if self.n > 0 {
             let pad = 8 - self.n;
-            self.put(((1u16 << pad) - 1) as u16, pad as u8);
+            self.put((1u16 << pad) - 1, pad as u8);
         }
     }
 }
